Adding validation, user roles and e-mail forgot password functionality

+++ database/migrations/2020_11_22_164713_create_coordinates_table.php
@@ -17,8 +17,8 @@ class CreateCoordinatesTable extends Migration
             $table->id()->autoIncrement();
             $table->foreignId('reservoir_id');
             $table->foreign('reservoir_id')->references('id')->on('reservoirs');
-            $table->decimal('lat');
-            $table->decimal('long');
+            $table->double('lat');
+            $table->double('long');
             $table->integer('radius');
             $table->timestamps();
         });
